cloud_kubernetes/server/requirements.txt
Flask>=2.1.1
numpy>=1.21.2
Pillow>=8.3.2
minio>=7.2.3
